+++ CHANGELOG.md
@@ -19,3 +19,11 @@
 - the parent directory is now configurable (instead of /opt)
 - the files in debian/skeleton will be copied: used for configuration files, documentation...
 
+## 1.0.4
+
+- the command understands now modes: 
+  - **help**: prints an usage message
+  - **create** : creates the Debian package
+  - **dependencies**: finds dependencies of library files
+- adapted README
+